pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "squix"
version = "0.1.0"
description = "AI Agent OS — orchestrate multiple AI models and agents as a managed team"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.9,<4",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "rich>=13.0",
    "httpx>=0.27,<1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
    "pre-commit",
]

[project.scripts]
squix = "squix.main:cli"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "SIM"]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.setuptools.packages.find]
include = ["squix*"]
